#579: hand judges the timeout they were given, not its floor
`build_options` called `to_i` on both values, so `--timeout 0.5` reached
a judge as `$options.timeout == 0` and `--lifetime 1.5` as `1`. The
enforcement above it does not truncate: `Timeout.timeout` and the
elapsed-time checks read `opts['lifetime']` directly. So a judge was told
one number while the runner acted on another, and the check in this very
method compared the floors rather than the values the caller passed.

`to_f` keeps them. The comparison below now sees the real numbers, which
is the point of having it.

Nothing existing changes: the only test that passes both values uses
lifetime 0.12 and timeout 0.1, and 0.12 < 0.11 is still false, so it
stays silent exactly as before.

The new test has a judge write `$options.timeout` into the factbase and
reads it back. Without this change it reads 0 instead of 0.5.

+++ lib/judges/commands/update.rb
@@ -71,7 +71,7 @@ class Judges::Update
   private
 
   def build_options(opts)
-    options = Judges::Options.new(timeout: opts['timeout']&.to_i, lifetime: opts['lifetime']&.to_i)
+    options = Judges::Options.new(timeout: opts['timeout']&.to_f, lifetime: opts['lifetime']&.to_f)
     if options.lifetime && options.timeout && options.lifetime < options.timeout * 1.1
       raise(
         StandardError,
